Seperate Comment componet from Comment List component
