+++ src/components/Comment/Comment.tsx
@@ -1,25 +1,19 @@
-import React, { useState, useEffect } from "react";
+import React from "react";
 import { css, cx } from "emotion";
 
 import { neutral, space, palette } from "@guardian/src-foundations";
 import { textSans } from "@guardian/src-foundations/typography";
 
-import { Pillar, CommentType, UserProfile, ThreadsType } from "../../types";
+import { Pillar, CommentType } from "../../types";
 import { GuardianStaff, GuardianPick } from "../Badges/Badges";
 import { RecommendationCount } from "../RecommendationCount/RecommendationCount";
 import { AbuseReportForm } from "../AbuseReportForm/AbuseReportForm";
 import { Timestamp } from "../Timestamp/Timestamp";
-import { CommentForm } from "../CommentForm/CommentForm";
-import { CommentReply } from "./CommentReply";
 
 type Props = {
   comment: CommentType;
   pillar: Pillar;
-  shortUrl: string;
-  user?: UserProfile;
-  onAddComment?: (commentId: number, body: string, user: UserProfile) => void;
-  displayReplyForm?: () => void;
-  threads: ThreadsType;
+  displayReplyForm: () => void;
 };
 
 const commentControls = css`
@@ -53,12 +47,6 @@ const commentWrapper = css`
   border-bottom: 1px solid ${neutral[86]};
   display: flex;
   padding: ${space[2]}px 0;
-`;
-
-const nestingStyles = css`
-  list-style-type: none;
-  padding-left: ${space[2]}px;
-  margin-left: ${space[12] + "px"};
 `;
 
 const commentAvatar = css`
@@ -96,42 +84,11 @@ const timestampWrapperStyles = css`
   justify-content: center;
 `;
 
-const buttonStyles = css`
-  margin-top: 12px;
-  margin-bottom: 12px;
-  cursor: pointer;
-  background: #fff;
-  color: #c70000;
-  height: 24px;
-  font-size: 12px;
-  font-weight: bold;
-  text-overflow: ellipsis;
-  border-radius: 12px;
-
-  border: 1px solid ${palette.neutral[86]};
-  svg {
-    fill: ${palette.neutral[60]};
-  }
-
-  :hover {
-    border: 1px solid ${palette.neutral[60]};
-    svg {
-      fill: ${palette.neutral[46]};
-    }
-  }
-`;
-
 export const avatar = (avatarSize: number): string => css`
   border-radius: ${avatarSize + 10}px;
   width: ${avatarSize}px;
   height: ${avatarSize}px;
 `;
-
-const Plus = () => (
-  <svg width="14" height="14" viewBox="0 0 18 18">
-    <path d="M8.2 0h1.6l.4 7.8 7.8.4v1.6l-7.8.4-.4 7.8H8.2l-.4-7.8L0 9.8V8.2l7.8-.4.4-7.8z"></path>
-  </svg>
-);
 
 const Column = ({ children }: { children: JSX.Element | JSX.Element[] }) => (
   <div
@@ -155,59 +112,8 @@ const Row = ({ children }: { children: JSX.Element | JSX.Element[] }) => (
   </div>
 );
 
-export const Comment = ({
-  comment,
-  pillar,
-  onAddComment,
-  user,
-  shortUrl,
-  threads
-}: Props) => {
+export const Comment = ({ comment, pillar, displayReplyForm }: Props) => {
   const commentControlsButtonStyles = commentControlsButton(pillar);
-
-  const [replyFormIsActive, setReplyFormIsActive] = useState<boolean>(false);
-  const [replyComment, setReplyComment] = useState<CommentType>(comment);
-  const displayReplyForm = () => setReplyFormIsActive(true);
-  const hideReplyForm = () => setReplyFormIsActive(false);
-
-  const displayCurrentReplyForm = () => {
-    setReplyComment(comment);
-    displayReplyForm();
-  };
-
-  // Filter logic
-  const [expanded, setExpanded] = useState<boolean>(threads === "expanded");
-  const [responses, setResponses] = useState(comment.responses);
-  const [loading, setLoading] = useState<boolean>(false);
-
-  const showResponses = threads !== "unthreaded";
-
-  const decideShowMoreText = () => {
-    const remainingResponses =
-      comment.metaData?.responseCount && comment.metaData?.responseCount - 3;
-    if (remainingResponses === 1) return `Show 1 more reply`;
-    return `Show ${remainingResponses} more replies`;
-  };
-
-  useEffect(() => {
-    setResponses(comment.responses);
-  }, [comment]);
-
-  const expand = (commentId: number) => {
-    setLoading(true);
-    fetch(
-      `http://discussion.code.dev-theguardian.com/discussion-api/comment/${commentId}?displayThreaded=true&displayResponses=true`
-    )
-      .then(response => response.json())
-      .then(json => {
-        setExpanded(true);
-        setResponses(json.comment.responses);
-      })
-      .finally(() => {
-        setLoading(false);
-      });
-  };
-
   return (
     <li>
       <div className={commentWrapper}>
@@ -255,7 +161,7 @@ export const Comment = ({
           <div className={spaceBetween}>
             <div className={commentControls}>
               <button
-                onClick={displayCurrentReplyForm}
+                onClick={displayReplyForm}
                 className={commentControlsButtonStyles}
               >
                 Reply
@@ -269,52 +175,6 @@ export const Comment = ({
           </div>
         </div>
       </div>
-
-      <>
-        {showResponses && responses && (
-          <div className={nestingStyles}>
-            {responses.map(comment => (
-              <CommentReply
-                comment={comment}
-                pillar={pillar}
-                displayReplyForm={() => {
-                  setReplyComment(comment);
-                  displayReplyForm();
-                }}
-              />
-            ))}
-            {!expanded && (
-              <button
-                onClick={() => expand(comment.id)}
-                className={buttonStyles}
-              >
-                <Row>
-                  <Plus />
-                  <span
-                    className={css`
-                      margin-left: 4px;
-                    `}
-                  >
-                    {loading ? "loading..." : decideShowMoreText()}
-                  </span>
-                </Row>
-              </button>
-            )}
-          </div>
-        )}
-        {replyFormIsActive && user && shortUrl && onAddComment && (
-          <div className={nestingStyles}>
-            <CommentForm
-              shortUrl={shortUrl}
-              onAddComment={onAddComment}
-              user={user}
-              hideReplyForm={hideReplyForm}
-              replyComment={replyComment}
-              defaultToActive={true}
-            />
-          </div>
-        )}
-      </>
     </li>
   );
 };
